formatting changes

+++ mgm_print_energy.h
@@ -5,6 +5,7 @@
 #define MGM_PRINT_ENERGY_H_
 /******************** PRINT ENERGY ********************/
 
+#include <cmath>
 #include "img.h"
 #include "img_tools.h"
 #include "point.h"
@@ -41,9 +42,9 @@ float evaluate_energy_4connected(const struct Img &u, std::vector<float > &outdi
       float Gtrunc=0; // contribution for the current point 
 
       // DATA TERM
-      int o = outdisp[pidx];
+      float o = outdisp[pidx];
       G      += CC[pidx][o];
-      GL2      += CC[pidx][o];
+      GL2    += CC[pidx][o];
       Gtrunc += CC[pidx][o];
 
       // EDGE POTENTIALS
@@ -59,8 +60,8 @@ float evaluate_energy_4connected(const struct Img &u, std::vector<float > &outdi
          if (!check_inside_image(pq ,u)) continue;
          int pridx = (pr.x+pr.y*nx);
          int pqidx = (pq.x+pq.y*nx);
-         int oor   = outdisp[pridx];
-         int ooq   = outdisp[pqidx];
+         float oor   = outdisp[pridx];
+         float ooq   = outdisp[pqidx];
 
          G   += fabs(oor - o)/N;
